Fixed random failures of zimcheck unit-tests
Before this fix, zimcheck's run-time reporting could be off by one
second. For example, if it took slightly less than 1 second, but started
and ended within the same second (thus not crossing the second boundary)
it would report a run time of 0 seconds. On the contrary, if it took a
small fraction of a second (e.g. 0.001 seconds) but started in second #i
and completed in second #(i+1) thus crossing the second boundary, then
it would report a run time of 1 second. Though not a big deal from the
user's perspective, the latter case resulted in infrequent failures of
the unit tests (that are very quick and contain a run time of 0 seconds
in their expected output). The fix is about calculating the run-time
more accurately (using `std::chrono::steady_clock`) and rounding it to
seconds, thus eliminating the effects of discreteness coming from coarse
second-level granularity.